% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updatehydat.R
\name{updatehydat}
\alias{updatehydat}
\title{Function to update provisional hydrometric data}
\usage{
updatehydat(station_number = names, parameters, ext = "*.rds")
}
\arguments{
\item{parameters}{Options are 46, (water level, primary sensor), 47 (discharge, primary sensory derived), 5 (water temperature), 16(water level, primary sensor), 52 (water level, primary sensor). List chosen parameters inside c().}
}
\value{
Saves new RDS files in the working directory.
}
\description{
Takes existing RDS files in the working directory and
updates them with real time hydrometric data.
}
\examples{
updatehydat(parameters = c(46, 47))
Saves new RDS files with updated water level and discharge values.
}
